🔨 refactor: components/common에 적절치 않았던 컴포넌트 옮기기
