+++ src/presentation/components/NeogaFormTicket/style.ts
