+++ src/presentation/components/NeogaResultCard/style.ts
