+++ src/presentation/components/NeogaResultComment/style.ts
